requirements.txt
pyftpdlib
pystray
Pillow
pywin32
pyinstaller